# backend/openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "info": {
    "title": "DSNP Gateway Prototype API Documentation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "authLogin",
        "summary": "Use Sign In With Frequency to login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletLoginRequest"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "202": {
            "description": "Account still pending creation"
          },
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletLoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "authLogout",
        "summary": "Logout and invalidate the access token",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "201": {
            "description": "Successful response"
          }
        }
      }
    },
    "/auth/account": {
      "get": {
        "operationId": "authAccount",
        "summary": "For polling to get the created account as authCreate can take time",
        "parameters": [
          {
            "name": "msaId",
            "in": "query",
            "description": "The user's Message Source Account ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "referenceId",
            "in": "query",
            "description": "A reference identifier to retrieve the resource when msaId is not available",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Account still pending creation"
          },
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthAccountResponse"
                }
              }
            }
          }
        }
      }
    },
    "/content/{dsnpId}": {
      "get": {
        "operationId": "getUserFeed",
        "summary": "Get recent posts from a user, paginated",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dsnpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newestBlockNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "oldestBlockNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBroadcast"
                }
              }
            }
          }
        }
      }
    },
    "/content/feed": {
      "get": {
        "operationId": "getFeed",
        "summary": "Get the Feed for the current user, paginated",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "newestBlockNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "oldestBlockNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBroadcast"
                }
              }
            }
          }
        }
      }
    },
    "/content/discover": {
      "get": {
        "operationId": "getDiscover",
        "summary": "Get the Discovery Feed for the current user, paginated",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "newestBlockNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "oldestBlockNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBroadcast"
                }
              }
            }
          }
        }
      }
    },
    "/content/create": {
      "post": {
        "operationId": "createBroadcast",
        "summary": "Create a new post",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastExtended"
                }
              }
            }
          }
        }
      }
    },
    "/content/{dsnpId}/{contentHash}": {
      "get": {
        "operationId": "getContent",
        "summary": "Get details of a specific post",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dsnpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contentHash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastExtended"
                }
              }
            }
          },
          "404": {
            "description": "Content not found"
          }
        }
      }
    },
    "/content/{type}/{contentHash}": {
      "put": {
        "operationId": "editContent",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "summary": "Edit the content of a specific post",
        "parameters": [
          {
            "name": "contentHash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "description": "Broadcast or Reply",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditPostRequest"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastExtended"
                }
              }
            }
          }
        }
      }
    },
    "/graph/{dsnpId}/following": {
      "get": {
        "operationId": "userFollowing",
        "summary": "Get a list of users that a specific user follows",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dsnpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "description": "Array of DSNP Ids",
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/graph/{dsnpId}/follow": {
      "post": {
        "operationId": "graphFollow",
        "summary": "Follow a user",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dsnpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "201": {
            "description": "Successful response"
          }
        }
      }
    },
    "/graph/{dsnpId}/unfollow": {
      "post": {
        "operationId": "graphUnfollow",
        "summary": "Unfollow a user",
        "security": [
          {
            "tokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "dsnpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "201": {
            "description": "Successful response"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "tokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "accessToken"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token invalid or not found"
      }
    },
    "schemas": {
...
      "WalletLoginResponse": {
        "type": "object",
        "properties": {
          "referenceId": {
            "type": "string"
          },
          "msaId": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          }
        },
        "required": [
          "referenceId"
        ]
      },
      "WalletLoginRequest": {
        "type": "object",
        "minProperties": 1,
        "maxProperties": 1,
        "additionalProperties": false,
        "properties": {
          "signIn": {
            "type": "object",
            "properties": {
              "siwsPayload": {
                "type": "object",
                "required": [
                  "message",
                  "signature"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  },
                  "signature": {
                    "type": "string"
                  }
                }
              },
              "error": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "signUp": {
            "type": "object",
            "properties": {
              "extrinsics": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": [
                    "pallet",
                    "extrinsicName",
                    "encodedExtrinsic"
                  ],
                  "properties": {
                    "pallet": {
                      "type": "string"
                    },
                    "extrinsicName": {
                      "type": "string"
                    },
                    "encodedExtrinsic": {
                      "type": "string"
                    }
                  }
                }
              },
              "error": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
...
      "AuthAccountResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expires": {
            "type": "integer"
          },
          "msaId": {
            "type": "string"
          },
          "displayHandle": {
            "type": "string"
          }
        },
        "required": [
          "msaId",
          "accessToken",
          "expires"
        ]
      },
...
      "PaginatedBroadcast": {
        "type": "object",
        "properties": {
          "newestBlockNumber": {
            "type": "integer"
          },
          "oldestBlockNumber": {
            "type": "integer"
          },
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BroadcastExtended"
            }
          }
        },
        "required": [
          "newestBlockNumber",
          "oldestBlockNumber",
          "posts"
        ]
      },
...
      "BroadcastExtended": {
        "type": "object",
        "properties": {
          "fromId": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          },
          "content": {
            "description": "JSON-encoded Activity Content Note",
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the post"
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReplyExtended"
            },
            "description": "Array of ReplyExtended objects"
          }
        },
        "required": [
          "fromId",
          "contentHash",
          "content",
          "timestamp"
        ]
      },
      "ReplyExtended": {
        "type": "object",
        "properties": {
          "fromId": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          },
          "content": {
            "description": "JSON-encoded Activity Content Note",
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the post"
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReplyExtended"
            },
            "description": "Array of ReplyExtended objects"
          }
        },
        "required": [
          "fromId",
          "contentHash",
          "content",
          "timestamp"
        ]
      },
...
      "CreatePostRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "inReplyTo": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            }
          }
        },
        "required": [
          "content"
        ]
      },
      "EditPostRequest": {
        "type": "object",
        "properties": {
          "targetContentHash": {
            "type": "string"
          },
          "targetAnnouncementType": {
            "type": "integer"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "content",
          "targetAnnouncementType",
          "targetContentHash"
        ]
      },
...
    }
  }
}